AI Demo 2.3/requirements.txt
# Root Cause AI System - Python Dependencies

# Core ML/AI Libraries
numpy==1.24.3
scikit-learn==1.3.0
networkx==3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Web Framework
flask==2.3.2
flask-cors==4.0.0

# Data Processing
pandas==2.0.3
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
